Moved waiting room analytics to the waiting room actions page
