+++ src/app/pages/waiting-room-to-car/waiting-room-to-car.analytics.effects.ts
@@ -1,16 +1,22 @@
-import {Injectable} from '@angular/core';
-import {CategoryCode} from '@dvsa/mes-test-schema/categories/common';
-import {Actions, createEffect, ofType} from '@ngrx/effects';
-import {select, Store} from '@ngrx/store';
-import {of} from 'rxjs';
-import {concatMap, filter, switchMap, withLatestFrom} from 'rxjs/operators';
+import { Injectable } from '@angular/core';
+import { CategoryCode } from '@dvsa/mes-test-schema/categories/common';
+import { Actions, createEffect, ofType } from '@ngrx/effects';
+import { Store, select } from '@ngrx/store';
+import { of } from 'rxjs';
+import { concatMap, filter, switchMap, withLatestFrom } from 'rxjs/operators';
 
-import {Router} from '@angular/router';
+import { Router } from '@angular/router';
+import { ModalEvent } from '@pages/waiting-room-to-car/components/mot-components/mot-failed-modal/mot-failed-modal.component';
+import { MOTAbortedMethod } from '@pages/waiting-room-to-car/components/vehicle-registration/vehicle-registration';
 import {
-  ModalEvent
-} from '@pages/waiting-room-to-car/components/mot-components/mot-failed-modal/mot-failed-modal.component';
-import {MOTAbortedMethod} from '@pages/waiting-room-to-car/components/vehicle-registration/vehicle-registration';
-import {
+  InvalidMotModalOutcome,
+  InvalidMotModalValidationError,
+  InvalidMotTerminate,
+  MotCallAborted,
+  MotFailedModalOpened,
+  MotNoEvidenceBannerCancelled,
+  MotSearchButtonPressed,
+  MotServiceUnavailable,
   WaitingRoomToCarBikeCategoryChanged,
   WaitingRoomToCarBikeCategorySelected,
   WaitingRoomToCarError,
@@ -18,8 +24,8 @@ import {
   WaitingRoomToCarViewBikeCategoryModal,
   WaitingRoomToCarViewDidEnter,
 } from '@pages/waiting-room-to-car/waiting-room-to-car.actions';
-import {AnalyticsProvider} from '@providers/analytics/analytics';
-import {AnalyticNotRecorded, AnalyticRecorded} from '@providers/analytics/analytics.actions';
+import { AnalyticsProvider } from '@providers/analytics/analytics';
+import { AnalyticNotRecorded, AnalyticRecorded } from '@providers/analytics/analytics.actions';
 import {
   AnalyticsScreenNames,
   GoogleAnalyticsCustomDimension,
@@ -27,21 +33,18 @@ import {
   GoogleAnalyticsEventsTitles,
   GoogleAnalyticsEventsValues,
 } from '@providers/analytics/analytics.model';
-import {AppConfigProvider} from '@providers/app-config/app-config';
-import {analyticsEventTypePrefix} from '@shared/helpers/format-analytics-text';
-import {StoreModel} from '@shared/models/store.model';
-import {getTestCategory} from '@store/tests/category/category.reducer';
-import {
-  getApplicationReference
-} from '@store/tests/journal-data/common/application-reference/application-reference.reducer';
-import {
-  getApplicationNumber
-} from '@store/tests/journal-data/common/application-reference/application-reference.selector';
-import {getCandidate} from '@store/tests/journal-data/common/candidate/candidate.reducer';
-import {getCandidateId} from '@store/tests/journal-data/common/candidate/candidate.selector';
-import {TestsModel} from '@store/tests/tests.model';
-import {getTests} from '@store/tests/tests.reducer';
-import {getCurrentTest, getJournalData, isPracticeMode} from '@store/tests/tests.selector';
+import { AppConfigProvider } from '@providers/app-config/app-config';
+import { MotStatusCodes } from '@providers/mot-history-api/mot-interfaces';
+import { analyticsEventTypePrefix } from '@shared/helpers/format-analytics-text';
+import { StoreModel } from '@shared/models/store.model';
+import { getTestCategory } from '@store/tests/category/category.reducer';
+import { getApplicationReference } from '@store/tests/journal-data/common/application-reference/application-reference.reducer';
+import { getApplicationNumber } from '@store/tests/journal-data/common/application-reference/application-reference.selector';
+import { getCandidate } from '@store/tests/journal-data/common/candidate/candidate.reducer';
+import { getCandidateId } from '@store/tests/journal-data/common/candidate/candidate.selector';
+import { TestsModel } from '@store/tests/tests.model';
+import { getTests } from '@store/tests/tests.reducer';
+import { getCurrentTest, getJournalData, isPracticeMode } from '@store/tests/tests.selector';
 import {
   OrditTrainedChanged,
   TrainerRegistrationNumberChanged,
@@ -54,31 +57,22 @@ import {
   PDILogbook,
   TraineeLicence,
 } from '@store/tests/trainer-details/cat-adi-part3/trainer-details.cat-adi-part3.actions';
-import {getTrainerDetails} from '@store/tests/trainer-details/cat-adi-part3/trainer-details.cat-adi-part3.reducer';
+import { getTrainerDetails } from '@store/tests/trainer-details/cat-adi-part3/trainer-details.cat-adi-part3.reducer';
 import {
   getPDILogbook,
   getTraineeLicence,
 } from '@store/tests/trainer-details/cat-adi-part3/trainer-details.cat-adi-part3.selector';
-import {getVehicleDetails} from '@store/tests/vehicle-details/cat-adi-part3/vehicle-details.cat-adi-part3.reducer';
-import {getDualControls} from '@store/tests/vehicle-details/cat-adi-part3/vehicle-details.cat-adi-part3.selector';
+import { getVehicleDetails } from '@store/tests/vehicle-details/cat-adi-part3/vehicle-details.cat-adi-part3.reducer';
+import { getDualControls } from '@store/tests/vehicle-details/cat-adi-part3/vehicle-details.cat-adi-part3.selector';
 import * as vehicleDetailsActions from '@store/tests/vehicle-details/vehicle-details.actions';
 import {
   DualControlsToggledNo,
   DualControlsToggledYes,
-  InvalidMotTerminate,
-  MotCallAborted,
   MotEvidenceProvidedToggled,
-  MotFailedModalOpened,
-  MotFailedModalOutcome,
-  MotFailedModalValidationError,
-  MotNoEvidenceBannerCancelled,
-  MotSearchButtonPressed,
-  MotServiceUnavailable,
   MotStatusChanged,
-  VehicleRegistrationChanged
+  VehicleRegistrationChanged,
 } from '@store/tests/vehicle-details/vehicle-details.actions';
-import {getMotStatus} from '@store/tests/vehicle-details/vehicle-details.selector';
-import {MotStatusCodes} from '@providers/mot-history-api/mot-interfaces';
+import { getMotStatus } from '@store/tests/vehicle-details/vehicle-details.selector';
 
 @Injectable()
 export class WaitingRoomToCarAnalyticsEffects {
@@ -585,7 +579,7 @@ export class WaitingRoomToCarAnalyticsEffects {
 
   motFailedModalOutcome$ = createEffect(() =>
     this.actions$.pipe(
-      ofType(MotFailedModalOutcome),
+      ofType(InvalidMotModalOutcome),
       concatMap((action) =>
         of(action).pipe(
           withLatestFrom(this.store$.pipe(select(getTests)), this.store$.pipe(select(getTests), select(isPracticeMode)))
@@ -594,7 +588,7 @@ export class WaitingRoomToCarAnalyticsEffects {
       filter(([, , practiceMode]) =>
         !practiceMode ? true : this.appConfigProvider.getAppConfig()?.journal?.enablePracticeModeAnalytics
       ),
-      switchMap(([{ modalEvent }, tests]: [ReturnType<typeof MotFailedModalOutcome>, TestsModel, boolean]) => {
+      switchMap(([{ modalEvent }, tests]: [ReturnType<typeof InvalidMotModalOutcome>, TestsModel, boolean]) => {
         // GA4 Analytics
         this.analytics.logGAEvent(
           analyticsEventTypePrefix(GoogleAnalyticsEvents.MOT_CHECK, tests),
@@ -681,7 +675,7 @@ export class WaitingRoomToCarAnalyticsEffects {
 
   motFailedModalValidationError$ = createEffect(() =>
     this.actions$.pipe(
-      ofType(MotFailedModalValidationError),
+      ofType(InvalidMotModalValidationError),
       concatMap((action) =>
         of(action).pipe(
           withLatestFrom(this.store$.pipe(select(getTests)), this.store$.pipe(select(getTests), select(isPracticeMode)))
@@ -690,7 +684,7 @@ export class WaitingRoomToCarAnalyticsEffects {
       filter(([, , practiceMode]) =>
         !practiceMode ? true : this.appConfigProvider.getAppConfig()?.journal?.enablePracticeModeAnalytics
       ),
-      switchMap(([, tests]: [ReturnType<typeof MotFailedModalValidationError>, TestsModel, boolean]) => {
+      switchMap(([, tests]: [ReturnType<typeof InvalidMotModalValidationError>, TestsModel, boolean]) => {
         // GA4 Analytics
         this.analytics.logGAEvent(
           analyticsEventTypePrefix(GoogleAnalyticsEvents.MOT_CHECK, tests),
