+++ src/store/tests/vehicle-details/cat-adi-part3/__tests__/vehicle-details.cat-adi-part3.reducer.spec.ts
@@ -1,6 +1,6 @@
+import { MotStatusCodes } from '@providers/mot-history-api/mot-interfaces';
 import { vehicleDetailsCatADIPart3Reducer } from '@store/tests/vehicle-details/cat-adi-part3/vehicle-details.cat-adi-part3.reducer';
 import * as vehicleDetailsActions from '../../vehicle-details.actions';
-import {MotStatusCodes} from '@providers/mot-history-api/mot-interfaces';
 
 describe('vehicleDetailsCatADIPart3Reducer', () => {
   describe('VehicleRegistrationChanged', () => {
