+++ src/store/tests/vehicle-details/cat-c/__tests__/vehicle-details.cat-c.reducer.spec.ts
@@ -1,7 +1,7 @@
 import { CatCUniqueTypes } from '@dvsa/mes-test-schema/categories/C';
+import { MotStatusCodes } from '@providers/mot-history-api/mot-interfaces';
 import { vehicleDetailsCatCReducer } from '@store/tests/vehicle-details/cat-c/vehicle-details.cat-c.reducer';
 import * as vehicleDetailsActions from '../../vehicle-details.actions';
-import {MotStatusCodes} from '@providers/mot-history-api/mot-interfaces';
 
 describe('vehicleDetailsCatDReducer', () => {
   describe('VehicleRegistrationChanged', () => {
@@ -15,7 +15,10 @@ describe('vehicleDetailsCatDReducer', () => {
   });
   describe('MotStatusChanged', () => {
     it('should set motStatus to the value given', () => {
-      const result = vehicleDetailsCatCReducer({ motStatus: null }, vehicleDetailsActions.MotStatusChanged(MotStatusCodes.NO_DETAILS));
+      const result = vehicleDetailsCatCReducer(
+        { motStatus: null },
+        vehicleDetailsActions.MotStatusChanged(MotStatusCodes.NO_DETAILS)
+      );
       expect(result).toEqual({ motStatus: MotStatusCodes.NO_DETAILS });
     });
   });
