+++ src/store/tests/vehicle-details/cat-d/__tests__/vehicle-details.cat-d.reducer.spec.ts
@@ -1,7 +1,7 @@
 import { CatDUniqueTypes } from '@dvsa/mes-test-schema/categories/D';
+import { MotStatusCodes } from '@providers/mot-history-api/mot-interfaces';
 import { vehicleDetailsCatDReducer } from '@store/tests/vehicle-details/cat-d/vehicle-details.cat-d.reducer';
 import * as vehicleDetailsActions from '../../vehicle-details.actions';
-import {MotStatusCodes} from '@providers/mot-history-api/mot-interfaces';
 
 describe('vehicleDetailsCatDReducer', () => {
   describe('VehicleRegistrationChanged', () => {
@@ -15,7 +15,10 @@ describe('vehicleDetailsCatDReducer', () => {
   });
   describe('MotStatusChanged', () => {
     it('should set motStatus to the value given', () => {
-      const result = vehicleDetailsCatDReducer({ motStatus: null }, vehicleDetailsActions.MotStatusChanged(MotStatusCodes.NO_DETAILS));
+      const result = vehicleDetailsCatDReducer(
+        { motStatus: null },
+        vehicleDetailsActions.MotStatusChanged(MotStatusCodes.NO_DETAILS)
+      );
       expect(result).toEqual({ motStatus: MotStatusCodes.NO_DETAILS });
     });
   });
